// server/src/configs/types.ts
export enum RequestType {
  LOGIN = 'LOGIN',
  REGISTER = 'REGISTER',
  CHECK_AUTH = 'CHECK_AUTH',
  FIND_USER = 'FIND_USER',
  GET_CONVERSATIONS = 'GET_CONVERSATIONS',
  GET_CONVERSATION_BY_ID = 'GET_CONVERSATION_BY_ID',
  SEND_TEXT_MESSAGE = 'SEND_TEXT_MESSAGE',
  SEND_FILE_MESSAGE = 'SEND_FILE_MESSAGE',
  NEW_PRIVATE_CONVERSATION = 'NEW_PRIVATE_CONVERSATION',
  NEW_GROUP_CONVERSATION = 'NEW_GROUP_CONVERSATION',
}

export enum ResponseType {
  LOGIN = 'LOGIN',
  REGISTER = 'REGISTER',
  CHECK_AUTH = 'CHECK_AUTH',
  FIND_USER = 'FIND_USER',
  GET_CONVERSATIONS = 'GET_CONVERSATIONS',
  GET_CONVERSATION_BY_ID = 'GET_CONVERSATION_BY_ID',
  NEW_MESSAGE = 'NEW_MESSAGE',
  READED_MESSAGE = 'READED_MESSAGE',
  NEW_PRIVATE_CONVERSATION = 'NEW_PRIVATE_CONVERSATION',
  NEW_GROUP_CONVERSATION = 'NEW_GROUP_CONVERSATION',
}

export enum ConversationType {
  PRIVATE = 'PRIVATE',
  GROUP = 'GROUP',
}

export enum MessageType {
  TEXT = 'TEXT',
  IMAGE = 'IMAGE',
  VIDEO = 'VIDEO',
}